Updates to the intro.
I could not read the comment Dr. Mac added to the first sentence about adding an "and ..." to the first sentence.
I can't tell what the second word he mean to write was, but other than that I think that I fixed all of the things that he included in his edits.